@@ -76,7 +76,7 @@ def test_der():
             raise ZeroDivisionError
 
     # Parse the outer element.
-    outer = reader.read_element(SEQUENCE)  # type: ignore[unreachable]
+    outer = reader.read_element(SEQUENCE)
     reader.check_empty()
     assert outer.data.tobytes() == der[2:]
 
