@@ -8,7 +8,7 @@ def prepare_doctest(x):
     return dedent(x).strip().splitlines(True)
 
 
-def fix_f2c_input(f2c_input_path: str):
+def fix_f2c_input(f2c_input_path: str) -> None:
     """
     CLAPACK has been manually modified to remove useless arguments generated by
     f2c. But the mismatches between the f2c ABI and the human-curated sensible
@@ -194,7 +194,7 @@ def char1_args_to_int(lines):
     return new_lines
 
 
-def fix_f2c_output(f2c_output_path: str):
+def fix_f2c_output(f2c_output_path: str) -> str | None:
     """
     This function is called on the name of each C output file. It fixes up the C
     output in various ways to compensate for the lack of f2c support for Fortran
@@ -271,8 +271,10 @@ def fix_f2c_output(f2c_output_path: str):
     with open(f2c_output, "w") as f:
         f.writelines(lines)
 
+    return None
 
-def add_externs_to_structs(lines: list[str]):
+
+def add_externs_to_structs(lines: list[str]) -> None:
     """
     The fortran "common" keyword is supposed to share variables between a bunch
     of files. f2c doesn't handle this correctly (it isn't possible for it to
