@@ -5,6 +5,7 @@ Build all of the packages in a given directory.
 """
 
 import argparse
+import dataclasses
 import hashlib
 import json
 import os
@@ -30,11 +31,13 @@ class BuildError(Exception):
         super().__init__()
 
 
+@total_ordering
+@dataclasses.dataclass(eq=False, repr=False)
 class BasePackage:
     pkgdir: Path
     name: str
     version: str
-    meta: dict
+    meta: dict[str, Any]
     library: bool
     shared_library: bool
     dependencies: list[str]
@@ -43,6 +46,7 @@ class BasePackage:
     unvendored_tests: Path | None = None
     file_name: str | None = None
     install_dir: str = "site"
+    _queue_idx: int | None = None
 
     # We use this in the priority queue, which pops off the smallest element.
     # So we want the smallest element to have the largest number of dependents
@@ -60,8 +64,11 @@ class BasePackage:
             self.pkgdir, self.pkgdir / "build", self.meta.get("source", {})
         )
 
+    def build(self, outputdir: Path, args: Any) -> None:
+        raise NotImplementedError()
 
-@total_ordering
+
+@dataclasses.dataclass
 class StdLibPackage(BasePackage):
     def __init__(self, pkgdir: Path):
         self.pkgdir = pkgdir
@@ -75,12 +82,18 @@ class StdLibPackage(BasePackage):
         self.dependents = set()
         self.install_dir = "lib"
 
-    def build(self, outputdir: Path, args) -> None:
+    def build(self, outputdir: Path, args: Any) -> None:
         # All build / packaging steps are already done in the main Makefile
         return
 
+    def wheel_path(self) -> Path:
+        raise RuntimeError("StdLibPackage has no wheel")
 
-@total_ordering
+    def tests_path(self) -> Path | None:
+        return None
+
+
+@dataclasses.dataclass
 class Package(BasePackage):
     def __init__(self, pkgdir: Path):
         self.pkgdir = pkgdir
@@ -108,7 +121,7 @@ class Package(BasePackage):
         dist_dir = self.pkgdir / "dist"
         wheels = list(find_matching_wheels(dist_dir.glob("*.whl")))
         if len(wheels) != 1:
-            raise Exception(
+            raise RuntimeError(
                 f"Unexpected number of wheels {len(wheels)} when building {self.name}"
             )
         return wheels[0]
@@ -120,7 +133,7 @@ class Package(BasePackage):
             return tests[0]
         return None
 
-    def build(self, outputdir: Path, args) -> None:
+    def build(self, outputdir: Path, args: Any) -> None:
 
         p = subprocess.run(
             [
@@ -174,11 +187,6 @@ class Package(BasePackage):
             shutil.copy(file_path, outputdir)
             file_path.unlink()
             return
-
-        shutil.copy(self.wheel_path(), outputdir)
-        test_path = self.tests_path()
-        if test_path:
-            shutil.copy(test_path, outputdir)
 
 
 def generate_dependency_graph(
@@ -245,14 +253,14 @@ def generate_dependency_graph(
     return pkg_map
 
 
-def job_priority(pkg: BasePackage):
+def job_priority(pkg: BasePackage) -> int:
     if pkg.name == "numpy":
         return 0
     else:
         return 1
 
 
-def print_with_progress_line(str, progress_line):
+def print_with_progress_line(str, progress_line) -> None:
     if not sys.stdout.isatty():
         print(str)
         return
@@ -263,7 +271,7 @@ def print_with_progress_line(str, progress_line):
         print(progress_line, end="\r")
 
 
-def get_progress_line(package_set):
+def get_progress_line(package_set) -> str | None:
     if not package_set:
         return None
     return "In progress: " + ", ".join(package_set.keys())
@@ -288,7 +296,7 @@ def format_name_list(l: list[str]) -> str:
 
 def mark_package_needs_build(
     pkg_map: dict[str, BasePackage], pkg: BasePackage, needs_build: set[str]
-):
+) -> None:
     """
     Helper for generate_needs_build_set. Modifies needs_build in place.
     Recursively add pkg and all of its dependencies to needs_build.
@@ -339,7 +347,7 @@ def build_from_graph(pkg_map: dict[str, BasePackage], outputdir: Path, args) -> 
 
     # Insert packages into build_queue. We *must* do this after counting
     # dependents, because the ordering ought not to change after insertion.
-    build_queue: PriorityQueue = PriorityQueue()
+    build_queue: PriorityQueue[tuple[int, BasePackage]] = PriorityQueue()
 
     if args.force_rebuild:
         # If "force_rebuild" is set, just rebuild everything
@@ -370,7 +378,7 @@ def build_from_graph(pkg_map: dict[str, BasePackage], outputdir: Path, args) -> 
         if len(pkg.unbuilt_dependencies) == 0:
             build_queue.put((job_priority(pkg), pkg))
 
-    built_queue: Queue = Queue()
+    built_queue: Queue[BasePackage | Exception] = Queue()
     thread_lock = Lock()
     queue_idx = 1
     # Using dict keys for insertion order preservation
@@ -411,11 +419,15 @@ def build_from_graph(pkg_map: dict[str, BasePackage], outputdir: Path, args) -> 
 
     num_built = len(already_built)
     while num_built < len(pkg_map):
-        pkg = built_queue.get()
-        if isinstance(pkg, BuildError):
-            raise SystemExit(pkg.returncode)
-        if isinstance(pkg, Exception):
-            raise pkg
+        match built_queue.get():
+            case BuildError() as err:
+                raise SystemExit(err.returncode)
+            case Exception() as err:
+                raise err
+            case a_package:
+                # MyPy should understand that this is a BasePackage
+                assert not isinstance(a_package, Exception)
+                pkg = a_package
 
         num_built += 1
 
@@ -439,7 +451,9 @@ def _generate_package_hash(full_path: Path) -> str:
     return sha256_hash.hexdigest()
 
 
-def generate_packages_json(output_dir: Path, pkg_map: dict[str, BasePackage]) -> dict:
+def generate_packages_json(
+    output_dir: Path, pkg_map: dict[str, BasePackage]
+) -> dict[str, dict[str, Any]]:
     """Generate the package.json file"""
     # Build package.json data.
     package_data: dict[str, dict[str, Any]] = {
@@ -452,12 +466,14 @@ def generate_packages_json(output_dir: Path, pkg_map: dict[str, BasePackage]) ->
     for name, pkg in pkg_map.items():
         if not pkg.file_name:
             continue
+        if not Path(output_dir, pkg.file_name).exists():
+            continue
         pkg_entry: Any = {
             "name": name,
             "version": pkg.version,
             "file_name": pkg.file_name,
             "install_dir": pkg.install_dir,
-            "sha_256": _generate_package_hash(Path(output_dir, pkg.file_name)),
+            "sha256": _generate_package_hash(Path(output_dir, pkg.file_name)),
         }
         if pkg.shared_library:
             pkg_entry["shared_library"] = True
@@ -479,6 +495,9 @@ def generate_packages_json(output_dir: Path, pkg_map: dict[str, BasePackage]) ->
                 "imports": [],
                 "file_name": pkg.unvendored_tests.name,
                 "install_dir": pkg.install_dir,
+                "sha256": _generate_package_hash(
+                    Path(output_dir, pkg.unvendored_tests.name)
+                ),
             }
             package_data["packages"][name.lower() + "-tests"] = pkg_entry
 
@@ -491,6 +510,18 @@ def generate_packages_json(output_dir: Path, pkg_map: dict[str, BasePackage]) ->
     package_data["packages"] = dict(sorted(package_data["packages"].items()))
 
     return package_data
+
+
+def copy_packages_to_dist_dir(packages, output_dir):
+    for pkg in packages:
+        try:
+            shutil.copy(pkg.wheel_path(), output_dir)
+        except RuntimeError:
+            pass
+
+        test_path = pkg.tests_path()
+        if test_path:
+            shutil.copy(test_path, output_dir)
 
 
 def build_packages(packages_dir: Path, output_dir: Path, args) -> None:
@@ -514,6 +545,7 @@ def build_packages(packages_dir: Path, output_dir: Path, args) -> None:
         pkg.file_name = pkg.wheel_path().name
         pkg.unvendored_tests = pkg.tests_path()
 
+    copy_packages_to_dist_dir(pkg_map.values(), output_dir)
     package_data = generate_packages_json(output_dir, pkg_map)
 
     with open(output_dir / "packages.json", "w") as fd:
@@ -532,12 +564,14 @@ def make_parser(parser):
         "dir",
         type=str,
         nargs=1,
+        default="packages",
         help="Input directory containing a tree of package definitions",
     )
     parser.add_argument(
         "output",
         type=str,
         nargs=1,
+        default="dist",
         help="Output directory in which to put all built packages",
     )
     parser.add_argument(
@@ -610,6 +644,7 @@ def make_parser(parser):
 def main(args):
     packages_dir = Path(args.dir[0]).resolve()
     outputdir = Path(args.output[0]).resolve()
+    outputdir.mkdir(exist_ok=True)
     if args.cflags is None:
         args.cflags = common.get_make_flag("SIDE_MODULE_CFLAGS")
     if args.cxxflags is None:
