@@ -1,6 +1,7 @@
 import functools
 import os
 import subprocess
+import sys
 from pathlib import Path
 from typing import Iterable, Iterator
 
@@ -8,7 +9,13 @@ import tomli
 from packaging.tags import Tag, compatible_tags, cpython_tags
 from packaging.utils import parse_wheel_filename
 
-PLATFORM = "emscripten_wasm32"
+from .io import parse_package_config
+
+
+def platform():
+    emscripten_version = get_make_flag("PYODIDE_EMSCRIPTEN_VERSION")
+    version = emscripten_version.replace(".", "_")
+    return f"emscripten_{version}_wasm32"
 
 
 def pyodide_tags() -> Iterator[Tag]:
@@ -19,6 +26,7 @@ def pyodide_tags() -> Iterator[Tag]:
     """
     PYMAJOR = get_make_flag("PYMAJOR")
     PYMINOR = get_make_flag("PYMINOR")
+    PLATFORM = platform()
     python_version = (int(PYMAJOR), int(PYMINOR))
     yield from cpython_tags(platforms=[PLATFORM], python_version=python_version)
     yield from compatible_tags(platforms=[PLATFORM], python_version=python_version)
@@ -154,11 +162,7 @@ def get_make_environment_vars():
 
     This allows us to set all build vars in one place"""
 
-    if "PYODIDE_ROOT" in os.environ:
-        PYODIDE_ROOT = Path(os.environ["PYODIDE_ROOT"])
-    else:
-        PYODIDE_ROOT = search_pyodide_root(os.getcwd())
-
+    PYODIDE_ROOT = get_pyodide_root()
     environment = {}
     result = subprocess.run(
         ["make", "-f", str(PYODIDE_ROOT / "Makefile.envs"), ".output_vars"],
@@ -203,3 +207,46 @@ def search_pyodide_root(curdir: str | Path, *, max_depth: int = 5) -> Path:
     raise FileNotFoundError(
         "Could not find Pyodide root directory. If you are not in the Pyodide directory, set `PYODIDE_ROOT=<pyodide-root-directory>`."
     )
+
+
+def init_environment():
+    if os.environ.get("__LOADED_PYODIDE_ENV"):
+        return
+    os.environ["__LOADED_PYODIDE_ENV"] = "1"
+    # If we are building docs, we don't need to know the PYODIDE_ROOT
+    if "sphinx" in sys.modules:
+        os.environ["PYODIDE_ROOT"] = ""
+
+    if "PYODIDE_ROOT" not in os.environ:
+        os.environ["PYODIDE_ROOT"] = str(search_pyodide_root(os.getcwd()))
+
+    os.environ.update(get_make_environment_vars())
+    hostsitepackages = get_hostsitepackages()
+    pythonpath = [
+        hostsitepackages,
+    ]
+    os.environ["PYTHONPATH"] = ":".join(pythonpath)
+    os.environ["BASH_ENV"] = ""
+    get_unisolated_packages()
+
+
+@functools.cache
+def get_pyodide_root():
+    init_environment()
+    return Path(os.environ["PYODIDE_ROOT"])
+
+
+@functools.cache
+def get_unisolated_packages():
+    import json
+
+    if "UNISOLATED_PACKAGES" in os.environ:
+        return json.loads(os.environ["UNISOLATED_PACKAGES"])
+    PYODIDE_ROOT = get_pyodide_root()
+    unisolated_packages = []
+    for pkg in (PYODIDE_ROOT / "packages").glob("**/meta.yaml"):
+        config = parse_package_config(pkg, check=False)
+        if config.get("build", {}).get("cross-build-env", False):
+            unisolated_packages.append(config["package"]["name"])
+    os.environ["UNISOLATED_PACKAGES"] = json.dumps(unisolated_packages)
+    return unisolated_packages
