@@ -27,6 +27,8 @@ PACKAGE_CONFIG_SPEC: dict[str, dict[str, Any]] = {
         "post": str,
         "replace-libs": list,
         "unvendor-tests": bool,
+        "cross-build-env": bool,
+        "cross-build-files": list,  # list[str]
     },
     "requirements": {
         "run": list,  # List[str],
@@ -150,7 +152,7 @@ def _check_config_wheel_build(config: dict[str, Any]) -> Iterator[str]:
     if "build" not in config:
         return
     build_metadata = config["build"]
-    allowed_keys = {"post", "unvendor-tests"}
+    allowed_keys = {"post", "unvendor-tests", "cross-build-env", "cross-build-files"}
     for key in build_metadata.keys():
         if key not in PACKAGE_CONFIG_SPEC["build"]:
             continue
@@ -186,7 +188,9 @@ def check_package_config_generate_errors(
     yield from _check_config_wheel_build(config)
 
 
-def check_package_config(config: dict[str, Any], file_path: Path | str | None = None):
+def check_package_config(
+    config: dict[str, Any], file_path: Path | str | None = None
+) -> None:
     errors_msg = list(check_package_config_generate_errors(config))
 
     if errors_msg:
@@ -197,7 +201,7 @@ def check_package_config(config: dict[str, Any], file_path: Path | str | None = 
         )
 
 
-def parse_package_config(path: Path | str, check: bool = True) -> dict[str, Any]:
+def parse_package_config(path: Path | str, *, check: bool = True) -> dict[str, Any]:
     """Load a meta.yaml file
 
     Parameters
