@@ -4,30 +4,28 @@ import sys
 import traceback
 from itertools import chain
 from pathlib import Path
-from typing import Mapping
+from typing import Generator, Mapping
 
-from build import BuildBackendException, ProjectBuilder  # type: ignore[import]
-from build.__main__ import (  # type: ignore[import]
+from build import BuildBackendException, ProjectBuilder
+from build.__main__ import (
     _STYLES,
     _error,
     _handle_build_error,
     _IsolatedEnvBuilder,
     _ProjectBuilder,
 )
-from build.env import IsolatedEnv  # type: ignore[import]
+from build.env import IsolatedEnv
 from packaging.requirements import Requirement
 
-from .common import get_hostsitepackages, get_pyversion
-
-UNISOLATED_PACKAGES = ["numpy", "scipy", "cffi", "pycparser", "pythran"]
+from .common import get_hostsitepackages, get_pyversion, get_unisolated_packages
 
 
-def symlink_unisolated_packages(env: IsolatedEnv):
+def symlink_unisolated_packages(env: IsolatedEnv) -> None:
     pyversion = get_pyversion()
     site_packages_path = f"lib/{pyversion}/site-packages"
-    env_site_packages = Path(env._path) / site_packages_path
+    env_site_packages = Path(env.path) / site_packages_path  # type: ignore[attr-defined]
     host_site_packages = Path(get_hostsitepackages())
-    for name in UNISOLATED_PACKAGES:
+    for name in get_unisolated_packages():
         for path in chain(
             host_site_packages.glob(f"{name}*"), host_site_packages.glob(f"_{name}*")
         ):
@@ -38,14 +36,14 @@ def symlink_unisolated_packages(env: IsolatedEnv):
 def remove_unisolated_requirements(requires: set[str]) -> set[str]:
     for reqstr in list(requires):
         req = Requirement(reqstr)
-        for avoid_name in UNISOLATED_PACKAGES:
+        for avoid_name in get_unisolated_packages():
             if avoid_name in req.name.lower():
                 requires.remove(reqstr)
     return requires
 
 
 @contextlib.contextmanager
-def replace_env(build_env: Mapping[str, str]):
+def replace_env(build_env: Mapping[str, str]) -> Generator[None, None, None]:
     old_environ = dict(os.environ)
     os.environ.clear()
     os.environ.update(build_env)
@@ -56,13 +54,13 @@ def replace_env(build_env: Mapping[str, str]):
         os.environ.update(old_environ)
 
 
-def install_reqs(env: IsolatedEnv, reqs: set[str]):
+def install_reqs(env: IsolatedEnv, reqs: set[str]) -> None:
     env.install(remove_unisolated_requirements(reqs))
     # Some packages (numcodecs) don't declare cython as a build dependency and
     # only recythonize if it is present. We need them to always recythonize so
     # we always install cython. If the reqs included some cython version already
     # then this won't do anything.
-    env.install(["cython"])
+    env.install(["cython", "pythran"])
 
 
 def _build_in_isolated_env(
@@ -99,7 +97,7 @@ def _build_in_isolated_env(
 def build(build_env: Mapping[str, str]):
     srcdir = Path.cwd()
     outdir = srcdir / "dist"
-    builder = _ProjectBuilder(srcdir)
+    builder = _ProjectBuilder(str(srcdir))
     distribution = "wheel"
     try:
         with _handle_build_error():
