@@ -1,6 +1,7 @@
 from collections import namedtuple
 from pathlib import Path
 from time import sleep
+from typing import Any
 
 import pytest
 
@@ -47,7 +48,7 @@ def test_generate_packages_json(tmp_path):
         "depends": ["pkg_1_1", "pkg_3"],
         "imports": ["pkg_1"],
         "install_dir": "site",
-        "sha_256": "c1e38241013b5663e902fff97eb8585e98e6df446585da1dcf2ad121b52c2143",
+        "sha256": "c1e38241013b5663e902fff97eb8585e98e6df446585da1dcf2ad121b52c2143",
     }
 
 
@@ -56,7 +57,7 @@ def test_build_dependencies(n_jobs, monkeypatch):
     build_list = []
 
     class MockPackage(buildall.Package):
-        def build(self, outputdir: Path, args) -> None:
+        def build(self, outputdir: Path, args: Any) -> None:
             build_list.append(self.name)
 
     monkeypatch.setattr(buildall, "Package", MockPackage)
@@ -87,7 +88,7 @@ def test_build_all_dependencies(n_jobs, monkeypatch):
     class MockPackage(buildall.Package):
         n_builds = 0
 
-        def build(self, outputdir: Path, args) -> None:
+        def build(self, outputdir: Path, args: Any) -> None:
             sleep(0.005)
             self.n_builds += 1
             # check that each build is only run once
@@ -108,7 +109,7 @@ def test_build_error(n_jobs, monkeypatch):
     """Try building all the dependency graph, without the actual build operations"""
 
     class MockPackage(buildall.Package):
-        def build(self, outputdir: Path, args) -> None:
+        def build(self, outputdir: Path, args: Any) -> None:
             raise ValueError("Failed build")
 
     monkeypatch.setattr(buildall, "Package", MockPackage)
