@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from typing import Any
 
 import pytest
 
@@ -41,7 +42,7 @@ def _args_wrapper(func):
 f2c_wrap = _args_wrapper(replay_f2c)
 
 
-def generate_args(line: str, args, is_link_cmd=False) -> str:
+def generate_args(line: str, args: Any, is_link_cmd: bool = False) -> str:
     splitline = line.split()
     res = handle_command_generate_args(splitline, args, is_link_cmd)
     for arg in [
