@@ -64,8 +64,10 @@ declare var globalThis: any;
 if (globalThis.FinalizationRegistry) {
   Module.finalizationRegistry = new FinalizationRegistry(
     ([ptr, cache]: [ptr: number, cache: PyProxyCache]) => {
-      cache.leaked = true;
-      pyproxy_decref_cache(cache);
+      if (cache) {
+        cache.leaked = true;
+        pyproxy_decref_cache(cache);
+      }
       try {
         Module._Py_DecRef(ptr);
       } catch (e) {
