@@ -9,7 +9,7 @@ try:
 except ImportError:
     pass
 
-from ._core import IN_BROWSER
+from ._core import IN_BROWSER, JsException
 from ._package_loader import unpack_buffer
 
 __all__ = [
@@ -130,7 +130,7 @@ class FetchResponse:
         self._raise_if_failed()
         return await self.js_response.text()
 
-    async def json(self, **kwargs) -> Any:
+    async def json(self, **kwargs: Any) -> Any:
         """Return the response body as a Javascript JSON object.
 
         Any keyword arguments are passed to `json.loads
@@ -149,7 +149,7 @@ class FetchResponse:
         self._raise_if_failed()
         return (await self.buffer()).to_bytes()
 
-    async def _into_file(self, f: TextIO | BinaryIO):
+    async def _into_file(self, f: TextIO | BinaryIO) -> None:
         """Write the data into an empty file with no copy.
 
         Warning: should only be used when f is an empty file, otherwise it may
@@ -158,7 +158,7 @@ class FetchResponse:
         buf = await self.buffer()
         buf._into_file(f)
 
-    async def _create_file(self, path: str):
+    async def _create_file(self, path: str) -> None:
         """Uses the data to back a new file without copying it.
 
         This method avoids copying the data when creating a new file. If you
@@ -205,7 +205,7 @@ class FetchResponse:
         unpack_buffer(buf, filename=filename, format=format, extract_dir=extract_dir)
 
 
-async def pyfetch(url: str, **kwargs) -> FetchResponse:
+async def pyfetch(url: str, **kwargs: Any) -> FetchResponse:
     r"""Fetch the url and return the response.
 
     This functions provides a similar API to the JavaScript `fetch function
@@ -227,6 +227,9 @@ async def pyfetch(url: str, **kwargs) -> FetchResponse:
         from js import Object
         from js import fetch as _jsfetch
 
-    return FetchResponse(
-        url, await _jsfetch(url, to_js(kwargs, dict_converter=Object.fromEntries))
-    )
+    try:
+        return FetchResponse(
+            url, await _jsfetch(url, to_js(kwargs, dict_converter=Object.fromEntries))
+        )
+    except JsException as e:
+        raise OSError(e.js_error.message) from None
