@@ -19,7 +19,9 @@ def filter_info(info: dict[str, Any], browser: str) -> dict[str, Any]:
     return result
 
 
-def possibly_skip_test(request, info: dict[str, Any]) -> dict[str, Any]:
+def possibly_skip_test(
+    request: pytest.FixtureRequest, info: dict[str, Any]
+) -> dict[str, Any]:
     if "segfault" in info:
         pytest.skip(f"known segfault: {info['segfault']}")
 
@@ -73,7 +75,7 @@ def test_cpython_core(main_test, selenium, request):
         raise
 
 
-def get_test_info(test) -> tuple[str, dict[str, Any]]:
+def get_test_info(test: dict[str, Any] | str) -> tuple[str, dict[str, Any]]:
     if isinstance(test, dict):
         (name, info) = next(iter(test.items()))
     else:
